Implement Linux and BSD inspired advanced drivers subsystem
- Implemented `EvdevInputDevice` for Linux evdev multi-touch tracking and force feedback event streaming.
- Implemented `FreeBsdDrmConnector` for FreeBSD DRM/KMS atomic state commits and VBlank sync tracking.
- Implemented `OpenBsdDriverPledge` for OpenBSD pledge/unveil sandboxed driver MMIO/DMA boundary enforcement.
- Implemented `NetBsdRumpDriverHost` for NetBSD rump-kernel virtualized driver host and userspace hypercalls.
- Implemented `LinuxUrbQueue` for Linux USB Request Block (URB) transfer queues with async completion handling.
- Registered and re-exported driver abstractions in `src/drivers/mod.rs` and `src/lib.rs`.

+++ src/network/tcp_udp.rs
@@ -178,7 +178,7 @@ impl TCPConnection for SimpleSocket {
         Ok(())
     }
     fn get_state(&self) -> TCPState {
-        unsafe { core::mem::transmute(self.state.load(Ordering::SeqCst) as u32) }
+        unsafe { core::mem::transmute(self.state.load(Ordering::SeqCst)) }
     }
 }
